refactor: standardize component import paths for consistency
- Updated import paths for various components across the application to follow a consistent naming convention, enhancing clarity and maintainability.
- Renamed components to use lowercase in their import paths, including shared components, sections, and SEO-related components.
- Improved organization of imports in layout and page files to streamline the codebase and facilitate easier navigation.
