+++ apps/web/app/blog/[slug]/page.tsx
@@ -5,12 +5,12 @@ import Link from 'next/link'
 import { notFound } from 'next/navigation'
 import { cache } from 'react'
 
-import { ContainerLayout } from '@/components/ContainerLayout.component'
-import { BlogCTA } from '@/components/blog/BlogCTA.component'
-import { PostMarkdown } from '@/components/blog/PostMarkdown.component'
-import { RelatedPosts } from '@/components/blog/RelatedPosts.component'
-import { TableOfContents } from '@/components/blog/TableOfContents.component'
-import { Breadcrumbs } from '@/components/shared/Breadcrumbs.component'
+import { ContainerLayout } from '@/components/container-layout.component'
+import { BlogCTA } from '@/components/blog/blog-cta.component'
+import { PostMarkdown } from '@/components/blog/post-markdown.component'
+import { RelatedPosts } from '@/components/blog/related-posts.component'
+import { TableOfContents } from '@/components/blog/table-of-contents.component'
+import { Breadcrumbs } from '@/components/shared/breadcrumbs.component'
 import { getPublishedPostBySlug } from '@/lib/queries/blog/post-detail.query'
 import { getRelatedPosts } from '@/lib/queries/blog/related-posts.query'
 import { seoConfig } from '@/lib/seo-config'
