+++ apps/web/app/blog/page.tsx
@@ -2,10 +2,10 @@ import { WebPageSchema } from '@workspace/seo/react'
 import { BookOpen, Tag } from 'lucide-react'
 import type { Metadata } from 'next'
 
-import { ContainerLayout } from '@/components/ContainerLayout.component'
-import { BlogHeroSection } from '@/components/blog/BlogHeroSection.component'
-import { InfinitePostList } from '@/components/blog/InfinitePostList.component'
-import { Breadcrumbs } from '@/components/shared/Breadcrumbs.component'
+import { ContainerLayout } from '@/components/container-layout.component'
+import { BlogHeroSection } from '@/components/blog/blog-hero-section.component'
+import { InfinitePostList } from '@/components/blog/infinite-post-list.component'
+import { Breadcrumbs } from '@/components/shared/breadcrumbs.component'
 import { getPublishedPostCardsPage } from '@/lib/queries/blog/post-list.query'
 import { seoConfig } from '@/lib/seo-config'
 import { toNextMetadata } from '@/lib/seo/metadata'
