+++ apps/web/app/blog/tags/[slug]/page.tsx
@@ -2,9 +2,9 @@ import { WebPageSchema } from '@workspace/seo/react'
 import type { Metadata } from 'next'
 import { cache } from 'react'
 
-import { ContainerLayout } from '@/components/ContainerLayout.component'
-import { InfinitePostList } from '@/components/blog/InfinitePostList.component'
-import { Breadcrumbs } from '@/components/shared/Breadcrumbs.component'
+import { ContainerLayout } from '@/components/container-layout.component'
+import { InfinitePostList } from '@/components/blog/infinite-post-list.component'
+import { Breadcrumbs } from '@/components/shared/breadcrumbs.component'
 import { getPublishedPostCardsPage } from '@/lib/queries/blog/post-list.query'
 import { getActiveTagBySlug } from '@/lib/queries/blog/taxonomy.query'
 import { seoConfig } from '@/lib/seo-config'
