+++ apps/web/app/blog/tags/page.tsx
@@ -3,9 +3,9 @@ import { Hash, Tag, Zap } from 'lucide-react'
 import type { Metadata } from 'next'
 import Link from 'next/link'
 
-import { ContainerLayout } from '@/components/ContainerLayout.component'
-import { BlogHeroSection } from '@/components/blog/BlogHeroSection.component'
-import { Breadcrumbs } from '@/components/shared/Breadcrumbs.component'
+import { ContainerLayout } from '@/components/container-layout.component'
+import { BlogHeroSection } from '@/components/blog/blog-hero-section.component'
+import { Breadcrumbs } from '@/components/shared/breadcrumbs.component'
 import { listActiveTagsWithCounts } from '@/lib/queries/blog/taxonomy.query'
 import { seoConfig } from '@/lib/seo-config'
 import { toNextMetadata } from '@/lib/seo/metadata'
