+++ apps/web/app/layout.tsx
@@ -2,14 +2,14 @@ import { OrganizationSchema, WebSiteSchema } from '@workspace/seo/react'
 import '@workspace/ui/globals.css'
 import { Geist, Geist_Mono } from 'next/font/google'
 
-import { AnalyticsProvider } from '@/components/analytics/AnalyticsProvider.component'
+import { AnalyticsProvider } from '@/components/analytics/analytics-provider.component'
 import { PageViewTracker } from '@/components/analytics/page-view-tracker.component'
 import { ScrollDepthTracker } from '@/components/analytics/scroll-depth-tracker.component'
-import { CookieBanner } from '@/components/CookieBanner.component'
-import { Footer } from '@/components/layout/Footer.component'
-import { Header } from '@/components/layout/Header.component'
+import { CookieBanner } from '@/components/cookie-banner.component'
+import { Footer } from '@/components/layout/footer.component'
+import { Header } from '@/components/layout/header.component'
 import { Providers } from '@/components/providers'
-import { MobileCallButton } from '@/components/shared/MobileCallButton.component'
+import { MobileCallButton } from '@/components/shared/mobile-call-button.component'
 import { WebVitals } from '@/components/web-vitals.component'
 import { env } from '@/env'
 import { seoConfig } from '@/lib/seo-config'
