+++ apps/web/app/services/[slug]/page.tsx
@@ -20,15 +20,15 @@ import { BreadcrumbSchema, JsonLd, WebPageSchema } from '@workspace/seo/react'
 import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 
-import { ServiceBenefits } from '@/components/services/ServiceBenefits.component'
-import { ServiceCTA } from '@/components/services/ServiceCTA.component'
-import { ServiceFeatures } from '@/components/services/ServiceFeatures.component'
-import { ServiceGallery } from '@/components/services/ServiceGallery.component'
-import { ServiceHero } from '@/components/services/ServiceHero.component'
-import { ServiceProcess } from '@/components/services/ServiceProcess.component'
-import { Breadcrumbs } from '@/components/shared/Breadcrumbs.component'
-import { ContentWrapper } from '@/components/shared/ContentWrapper.component'
-import { FAQComponent } from '@/components/shared/FAQ.component'
+import { ServiceBenefits } from '@/components/services/service-benefits.component'
+import { ServiceCTA } from '@/components/services/service-cta.component'
+import { ServiceFeatures } from '@/components/services/service-features.component'
+import { ServiceGallery } from '@/components/services/service-gallery.component'
+import { ServiceHero } from '@/components/services/service-hero.component'
+import { ServiceProcess } from '@/components/services/service-process.component'
+import { Breadcrumbs } from '@/components/shared/breadcrumbs.component'
+import { ContentWrapper } from '@/components/shared/content-wrapper.component'
+import { FAQComponent } from '@/components/shared/faq.component'
 import { siteConfig } from '@/lib/data/site-config'
 import {
     getPublishedServices,
