+++ apps/web/app/services/page.tsx
@@ -17,15 +17,15 @@
  */
 import { BreadcrumbSchema, JsonLd, WebPageSchema } from '@workspace/seo/react'
 
-import { ServiceCardHorizontal } from '@/components/services/ServiceCardHorizontal.component'
-import { ServicesContact } from '@/components/services/ServicesContact.component'
-import { ServicesTestimonials } from '@/components/services/ServicesTestimonials.component'
-import { Breadcrumbs } from '@/components/shared/Breadcrumbs.component'
-import { ContentWrapper } from '@/components/shared/ContentWrapper.component'
-import { FAQComponent } from '@/components/shared/FAQ.component'
-import { Gallery } from '@/components/shared/Gallery.component'
-import { SectionContainer } from '@/components/shared/SectionContainer.component'
-import { SectionHeader } from '@/components/shared/SectionHeader.component'
+import { ServiceCardHorizontal } from '@/components/services/service-card-horizontal.component'
+import { ServicesContact } from '@/components/services/services-contact.component'
+import { ServicesTestimonials } from '@/components/services/services-testimonials.component'
+import { Breadcrumbs } from '@/components/shared/breadcrumbs.component'
+import { ContentWrapper } from '@/components/shared/content-wrapper.component'
+import { FAQComponent } from '@/components/shared/faq.component'
+import { Gallery } from '@/components/shared/gallery.component'
+import { SectionContainer } from '@/components/shared/section-container.component'
+import { SectionHeader } from '@/components/shared/section-header.component'
 import { servicesFAQ } from '@/lib/data/services/services-page-content'
 import { siteConfig } from '@/lib/data/site-config'
 import { getPublishedServices } from '@/lib/queries/get-services.query'
