+++ apps/web/app/terms/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 
-import { LegalPageLayout } from '@/components/legal/LegalPageLayout.component'
+import { LegalPageLayout } from '@/components/legal/legal-page-layout.component'
 import { termsOfServiceContent } from '@/lib/data/legal/terms-of-service.content'
 
 export const metadata: Metadata = {
