+++ apps/web/components/sections/about/mission-section.component.tsx
@@ -20,10 +20,10 @@
  */
 import { cn } from '@workspace/ui/lib/utils'
 
-import { ContentWrapper } from '@/components/shared/ContentWrapper.component'
-import { IconCard } from '@/components/shared/IconCard.component'
-import { SectionContainer } from '@/components/shared/SectionContainer.component'
-import { SectionHeader } from '@/components/shared/SectionHeader.component'
+import { ContentWrapper } from '@/components/shared/content-wrapper.component'
+import { IconCard } from '@/components/shared/icon-card.component'
+import { SectionContainer } from '@/components/shared/section-container.component'
+import { SectionHeader } from '@/components/shared/section-header.component'
 import type { MissionSectionProps } from '@/lib/types/sections/mission-section.type'
 
 export function MissionSection({
