+++ apps/web/components/sections/about/story-section.component.tsx
@@ -17,9 +17,9 @@
 import { cn } from '@workspace/ui/lib/utils'
 import Image from 'next/image'
 
-import { ContentWrapper } from '@/components/shared/ContentWrapper.component'
-import { SectionContainer } from '@/components/shared/SectionContainer.component'
-import { SectionHeader } from '@/components/shared/SectionHeader.component'
+import { ContentWrapper } from '@/components/shared/content-wrapper.component'
+import { SectionContainer } from '@/components/shared/section-container.component'
+import { SectionHeader } from '@/components/shared/section-header.component'
 import type { StorySectionProps } from '@/lib/types/sections/story-section.type'
 
 export function StorySection({
