+++ apps/web/components/sections/about/team-section.component.tsx
@@ -26,9 +26,9 @@ import { Linkedin, Mail, Twitter } from 'lucide-react'
 import Image from 'next/image'
 import Link from 'next/link'
 
-import { ContentWrapper } from '@/components/shared/ContentWrapper.component'
-import { SectionContainer } from '@/components/shared/SectionContainer.component'
-import { SectionHeader } from '@/components/shared/SectionHeader.component'
+import { ContentWrapper } from '@/components/shared/content-wrapper.component'
+import { SectionContainer } from '@/components/shared/section-container.component'
+import { SectionHeader } from '@/components/shared/section-header.component'
 import type { TeamSectionProps } from '@/lib/types/sections/team-section.type'
 
 /**
