+++ apps/web/components/sections/home/services-preview-section.component.tsx
@@ -27,10 +27,10 @@ import { Button } from '@workspace/ui/components/button'
 import { ArrowRight } from 'lucide-react'
 import Link from 'next/link'
 
-import { ServiceCard } from '@/components/services/ServiceCard.component'
-import { ContentWrapper } from '@/components/shared/ContentWrapper.component'
-import { SectionContainer } from '@/components/shared/SectionContainer.component'
-import { SectionHeader } from '@/components/shared/SectionHeader.component'
+import { ServiceCard } from '@/components/services/service-card.component'
+import { ContentWrapper } from '@/components/shared/content-wrapper.component'
+import { SectionContainer } from '@/components/shared/section-container.component'
+import { SectionHeader } from '@/components/shared/section-header.component'
 import { getPublishedServices } from '@/lib/queries/get-services.query'
 import type { CommonSectionProps } from '@/lib/types/sections/section.type'
