+++ packages/seo/src/react/article/article-schema.component.tsx
@@ -1,7 +1,7 @@
 import type { Article, BlogPosting, WithContext } from 'schema-dts'
 
 import { buildArticleJsonLd } from '../../schemas/article.schema'
-import { JsonLd } from '../JsonLd.component'
+import { JsonLd } from '../json-ld.component'
 import type { ArticleSchemaProps } from '../../types/schema/article.type'
 
 export function ArticleSchema(props: ArticleSchemaProps) {
