+++ packages/seo/src/react/breadcrumb/breadcrumb-schema.component.tsx
@@ -2,7 +2,7 @@ import type { BreadcrumbList, WithContext } from 'schema-dts'
 
 import { buildBreadcrumbJsonLd } from '../../schemas/breadcrumb.schema'
 import type { BreadcrumbSchemaProps } from '../../types/schema/breadcrumb.type'
-import { JsonLd } from '../JsonLd.component'
+import { JsonLd } from '../json-ld.component'
 
 export function BreadcrumbSchema(props: BreadcrumbSchemaProps) {
     const data: WithContext<BreadcrumbList> = buildBreadcrumbJsonLd(props)
