+++ packages/seo/src/react/index.ts
@@ -1,10 +1,10 @@
-export { JsonLd } from './JsonLd.component'
-export { OrganizationSchema } from './organization/OrganizationSchema.component'
-export { WebSiteSchema } from './website/WebSiteSchema.component'
-export { WebPageSchema } from './webpage/WebPageSchema.component'
-export { ArticleSchema } from './article/ArticleSchema.component'
-export { BreadcrumbSchema } from './breadcrumb/BreadcrumbSchema.component'
-export { FAQSchema } from './faq/FAQSchema.component'
-export { ProductSchema } from './product/ProductSchema.component'
-export { ReviewSchema } from './review/ReviewSchema.component'
-export { LocalBusinessSchema } from './local-business/LocalBusinessSchema.component'
+export { JsonLd } from './json-ld.component'
+export { OrganizationSchema } from './organization/organization-schema.component'
+export { WebSiteSchema } from './website/web-site-schema.component'
+export { WebPageSchema } from './webpage/web-page-schema.component'
+export { ArticleSchema } from './article/article-schema.component'
+export { BreadcrumbSchema } from './breadcrumb/breadcrumb-schema.component'
+export { FAQSchema } from './faq/faq-schema.component'
+export { ProductSchema } from './product/product-schema.component'
+export { ReviewSchema } from './review/review-schema.component'
+export { LocalBusinessSchema } from './local-business/local-business-schema.component'
