+++ packages/seo/src/react/local-business/local-business-schema.component.tsx
@@ -2,7 +2,7 @@ import type { LocalBusiness, WithContext } from 'schema-dts'
 
 import { buildLocalBusinessJsonLd } from '../../schemas/local-business.schema'
 import type { LocalBusinessSchemaProps } from '../../types/schema/local-business.type'
-import { JsonLd } from '../JsonLd.component'
+import { JsonLd } from '../json-ld.component'
 
 export function LocalBusinessSchema(props: LocalBusinessSchemaProps) {
     const data: WithContext<LocalBusiness> = buildLocalBusinessJsonLd(props)
