+++ packages/seo/src/react/organization/organization-schema.component.tsx
@@ -2,7 +2,7 @@ import type { Organization, WithContext } from 'schema-dts'
 
 import { buildOrganizationJsonLd } from '../../schemas/organization.schema'
 import type { OrganizationSchemaProps } from '../../types/schema/organization.type'
-import { JsonLd } from '../JsonLd.component'
+import { JsonLd } from '../json-ld.component'
 
 export function OrganizationSchema(props: OrganizationSchemaProps) {
     const data: WithContext<Organization> = buildOrganizationJsonLd(props)
