+++ packages/seo/src/react/product/product-schema.component.tsx
@@ -2,7 +2,7 @@ import type { Product, WithContext } from 'schema-dts'
 
 import { buildProductJsonLd } from '../../schemas/product.schema'
 import type { ProductSchemaProps } from '../../types/schema/product.type'
-import { JsonLd } from '../JsonLd.component'
+import { JsonLd } from '../json-ld.component'
 
 export function ProductSchema(props: ProductSchemaProps) {
     const data: WithContext<Product> = buildProductJsonLd(props)
