+++ packages/seo/src/react/review/review-schema.component.tsx
@@ -2,7 +2,7 @@ import type { Review, WithContext } from 'schema-dts'
 
 import { buildReviewJsonLd } from '../../schemas/review.schema'
 import type { ReviewSchemaProps } from '../../types/schema/review.type'
-import { JsonLd } from '../JsonLd.component'
+import { JsonLd } from '../json-ld.component'
 
 export function ReviewSchema(props: ReviewSchemaProps) {
     const data: WithContext<Review> = buildReviewJsonLd(props)
